==> 01-clean_and_filter_trip_data.ipynb <==
import os
import pandas as pd
from datetime import time

def filter_eligible_trips(df):
  eligible_trips = pd.DataFrame()
  max_stop_sequence = df.loc[df['stop_sequence'].idxmax()]["stop_sequence"]
  df['arrival_time_iso'] = pd.to_datetime(df['arrival_time'], unit='s')
  df['arrival_time_iso'] = df['arrival_time_iso'].dt.tz_localize('UTC')
  df['arrival_time_local'] = df['arrival_time_iso'].dt.tz_convert('Europe/Stockholm')

  for trip_id, trip_group in df.groupby("trip_id"):
    first_stop = trip_group[trip_group['stop_sequence'] == 1]
    last_stop = trip_group[trip_group['stop_sequence'] == max_stop_sequence]

    if first_stop.shape[0] == 0 or last_stop.shape[0] == 0:
      print(f"Trip {trip_id} missing first or last stop", first_stop.shape[0], last_stop.shape[0])
      continue

    first_arrival = first_stop["arrival_time_local"].iloc[0] 
    last_arrival = last_stop["arrival_time_local"].iloc[0]
      
    first_cutoff = first_arrival.replace(hour=6, minute=30, second=0, microsecond=0)
    last_cutoff = last_arrival.replace(hour=9, minute=0, second=0, microsecond=0)

    if first_arrival > first_cutoff and last_arrival < last_cutoff:
      eligible_trips = pd.concat([eligible_trips, trip_group])
   
  return eligible_trips



def concatenate_all_raw(type_day_str):
    trips_snow_raw = []

    for filename in os.listdir("./../raw_data/"+type_day_str):
        print(filename)
        if filename.split("-")[3] == "arrivals":
            df_s = pd.read_csv(f"./../raw_data/{type_day_str}/{filename}")
            df_s['date'] = pd.to_datetime(df_s['arrival_time'], unit='s').dt.date
            trips_snow_raw.append(filter_eligible_trips(df_s))

    df_all_snow = pd.concat(trips_snow_raw, ignore_index=True)
    df_all_snow['trip_day_id'] = df_all_snow.groupby(['trip_id', 'date']).ngroup()
    
    return df_all_snow



def process_trip_groups(df):
    results = []
    missing_stop_dict = {} 

    grouped = df.groupby('trip_day_id')

    for trip_id, group in grouped:

        # get all unique busstops in df ('stop_sequence' or 'stop_id')
        unique_stops = sorted(group['stop_sequence'].unique())
        missing_stops = set(range(1, group['stop_sequence'].max())) - set(unique_stops)

        # add trip_id to proper bucket based on which stops are missing
        key = frozenset(missing_stops)
        if key not in missing_stop_dict:
            missing_stop_dict[key] = []
        missing_stop_dict[key].append(trip_id)

        # get all certain stops
        # arrival-wise
        certain_arriv = group[
            group.arrival_uncertainty == 0.0
        ].drop_duplicates(subset=['stop_sequence'], keep='first').sort_values(by='stop_sequence')

        # departure-wise
        certain_depart = group[
            (group.arrival_uncertainty != 0.0) & (group.departure_uncertainty == 0.0)
        ].drop_duplicates(subset=['stop_sequence'], keep='first').sort_values(by='stop_sequence')

        # find stops for which no certain info
        certain_stops = set(certain_arriv['stop_sequence']).union(set(certain_depart['stop_sequence']))
        uncertain_stops = set(unique_stops) - certain_stops

        # get best estimates (most recent)
        best_est = group[group['stop_sequence'].isin(uncertain_stops)].sort_values(by='update_timestamp', ascending=False).drop_duplicates(subset='stop_sequence', keep='first')

        # combine
        combined = pd.concat([certain_arriv, certain_depart, best_est], ignore_index=True).drop_duplicates(subset='stop_sequence', keep='first').sort_values(by='stop_sequence')

        results.append(combined)

    return pd.concat(results), missing_stop_dict

# get all raw_data & filter by time
trips_snow = concatenate_all_raw('snow_days')
trips_control = concatenate_all_raw('control_days')
trips_second_pull = concatenate_all_raw('second_pull')
trips_additional_control = concatenate_all_raw('additional_control')

dedup_snow, _ = process_trip_groups(trips_snow)
dedup_control, _ = process_trip_groups(trips_control)
dedup_second_pull, _ = process_trip_groups(trips_second_pull)
dedup_additional_control, _ = process_trip_groups(trips_additional_control)

weather_control = pd.read_csv('./control_days_DEPTH_BASED.csv')
weather_snow = pd.read_csv('./snow_days_DEPTH_BASED.csv')
weather_combined = pd.concat([weather_control, weather_snow], ignore_index=True)

trips_combined = pd.concat([dedup_snow, dedup_control, dedup_second_pull, dedup_additional_control], ignore_index=True)

trips_combined['date'] = pd.to_datetime(trips_combined['date'])
weather_combined['date'] = pd.to_datetime(weather_combined['date'])

all_trip_data = trips_combined[trips_combined['date'].isin(weather_combined['date'])]

all_trip_data

all_trip_data.to_csv("./all_trip_data.csv", index=False)
