==> 02-calc_daily_delay_metrics.ipynb <==
# Overview

Calculates three metrics at the daily level 
| Metric | Calculation | Captures | Ignores |
|--------|-------------|----------|---------|
| Mean Delay | simple avg of arrival delay of each stop in each trip -> avg of all trips for each day | Overall | correlation between delays for all stops in a trip, recovery |
| Net Additional Delay | `arrival_delay(n) - arrival_delay(n-1)` for each stop **->** avg of all stops in each trip **->** avg for all trips in a day | Schedule drift along a trip, or systematic lateness or earliness | Recovery behavior, delay at stop 1 |
| Absolute Additional Delay | `abs(net_additional_delay)` **->** avg of all stops in a trip **->** avg for all trips in a day | Early AND late arrivals | Net effect |
| Delay Recovery Index | `abs_additional_delay - abs(net_additional_dealy)` avg all trips in a day | Recovery: instability minus drift | final outcome |

# SETUP

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

# read trip and weather data 
control_weather = pd.read_csv('./control_days_DEPTH_BASED.csv')
snow_weather = pd.read_csv('./snow_days_DEPTH_BASED.csv')
trip_data = pd.read_csv('all_trip_data.csv')

trip_data

# Calculations


# stolen from Marina's notebook 🙏
def calc_additional_arrival_delay(df):
    # additional delay at each station (does not capture or penalize delay recovery or early arrivals)
    df['net_additional_arrival_delay'] = df.groupby('trip_day_id')['arrival_delay'].diff().fillna(0)
    df.loc[df['stop_sequence'] == 1, 'net_additional_arrival_delay'] = 0

    # absolute value of the net delay to "penalize" recovery
    df['abs_additional_arrival_delay'] = df['net_additional_arrival_delay'].abs()

    return df

calc_additional_arrival_delay(trip_data)
trip_data

# per-trip averages for each version of accumulated delay
trip_delay_metrics = trip_data.groupby(['date', 'trip_day_id']).agg({
    'arrival_delay': 'mean',
    'net_additional_arrival_delay': 'mean',
    'abs_additional_arrival_delay': 'mean',
    'stop_sequence': 'max'
}).reset_index().rename(columns={
    'arrival_delay': 'mean_arrival_delay',
    'net_additional_arrival_delay': 'mean_net_additional_delay',
    'abs_additional_arrival_delay': 'mean_abs_additional_delay',
    'stop_sequence': 'num_stops'
})

# "delay recovery index"
# low values -> delays kept accumulating and were not recovered (the trips finished with delay)
# high values -> within trip fluctuations were mostly recovered and trips tended to finish on time
trip_delay_metrics['delay_recovery_index'] = trip_delay_metrics['mean_abs_additional_delay'] - trip_delay_metrics['mean_net_additional_delay'].abs()



# aggregate to daily averages
daily_delay_metrics = trip_delay_metrics.groupby(['date']).agg({
    'mean_arrival_delay': 'mean',
    'mean_net_additional_delay': 'mean',
    'mean_abs_additional_delay': 'mean',
    'delay_recovery_index': 'mean'
}).reset_index().rename(columns={
    'delay_recovery_index': 'mean_delay_recovery_index'
})


daily_delay_metrics

# Merge weather with daily delay
snow_weather_with_delay = pd.merge(snow_weather, daily_delay_metrics, 'left', on="date").drop(["Unnamed: 0", "date_dt"], axis=1)
control_weather_with_delay = pd.merge(control_weather, daily_delay_metrics, 'left', on="date").drop(["Unnamed: 0", "date_dt"], axis=1)

snow_weather_with_delay = snow_weather_with_delay[['date'] + [col for col in snow_weather_with_delay.columns if col != 'date']]
control_weather_with_delay = control_weather_with_delay[['date'] + [col for col in control_weather_with_delay.columns if col != 'date']]

# write to file
snow_weather_with_delay.to_csv("./weather_delay_SNOW.csv", index=False)
control_weather_with_delay.to_csv("./weather_delay_CONTROL.csv", index=False)
